WIP: code refactored and moved usestate to hooks
